Refactor container_resolvers
